API

API

Types

Functions

BitOperations.bclear

BitOperations.bflip

BitOperations.bflip

BitOperations.bget

BitOperations.bget

BitOperations.bmask

BitOperations.bmask

BitOperations.bset

BitOperations.bset

BitOperations.bset

BitOperations.bsizeof

BitOperations.lsbget

BitOperations.lsbget

BitOperations.lsbmask

BitOperations.lsbmask

BitOperations.msbget

BitOperations.msbget

BitOperations.msbmask

BitOperations.msbmask

BitOperations.zigzagdec

BitOperations.zigzagenc

Documentation

BitOperations.bclear — Method.

bclear(x::T, bit::Integer)::T where {T<:Integer}

Returns a modified copy of x, with bit bit cleared (set to zero).

source

BitOperations.bflip — Method.

bflip(x::T, bit::Integer)::T where {T<:Integer}

Returns a modified copy of x, with bit bit flipped.

source

BitOperations.bflip — Method.

bflip(x::T, bits::UnitRange{<:Integer})::T where {T<:Integer}

Returns a modified copy of x, with all bits in bit range bits flipped.

source

BitOperations.bget — Method.

bget(x::T, bit::Integer)::Bool where {T<:Integer}

Get the value of bit bit of x.

source

BitOperations.bget — Method.

bget(x::T, bits::UnitRange{<:Integer})::T where {T<:Integer}

Get the value of the bit range bits of x.

source

BitOperations.bmask — Method.

bmask(::Type{T}, bit::Integer)::T where {T<:Integer}

Generates a bit mask of type T with bit bit set to one.

source

BitOperations.bmask — Method.

bmask(::Type{T}, bits::UnitRange{<:Integer})::T where {T<:Integer}

Generates a bit mask of type T with bit-range bits set to one.

source

BitOperations.bset — Method.

bset(x::T, bit::Integer, y::Bool)::T where {T<:Integer}

Returns a modified copy of x, with bit bit set to y.

source

BitOperations.bset — Method.

bset(x::T, bit::Integer)::T where {T<:Integer}

Returns a modified copy of x, with bit bit set (to one).

source

BitOperations.bset — Method.

bset(x::T, bits::UnitRange{<:Integer}, y::Integer)::T where {T<:Integer}

Returns a modified copy of x, with bit range bits set to y.

source

BitOperations.bsizeof — Method.

bsizeof(x)

Returns the data size of x in bits.

source

BitOperations.lsbget — Method.

lsbget(x::T)::T where {T <: Integer}

Returns the value of the nbits least significant bits of x.

source

BitOperations.lsbget — Method.

lsbget(x::Integer)::Bool

Returns the value of the least significant bit of x.

source

BitOperations.lsbmask — Method.

lsbmask(::Type{T})::T where {T<:Integer}

Generates a bit mask with only the least significant bit set.

source

BitOperations.lsbmask — Method.

lsbmask(::Type{T}, nbits::Integer)::T where {T<:Integer}

Generates a bit mask with only the nbits least significant bits set.

source

BitOperations.msbget — Method.

msbget(x::T)::T where {T<:Integer}

Returns the value of the nbits most significant bits of x.

source

BitOperations.msbget — Method.

msbget(x::T)::Bool where {T<:Integer}

Returns the value of the most significant bit of x.

source

BitOperations.msbmask — Method.

msbmask(::Type{T})::T where {T<:Integer}

Generates a bit mask with only the most significant bit set.

source

BitOperations.msbmask — Method.

msbmask(::Type{T}, nbits::Integer)::T where {T<:Integer}

Generates a bit mask with only the nbits most significant bits set.

source

BitOperations.zigzagdec — Function.

zigzagdec(x::Unsigned)::Signed

Zigzag-decode x, Google Protocol Buffers compatible.

source

BitOperations.zigzagenc — Function.

zigzagenc(x::Signed)::Unsigned

Zigzag-encode x, Google Protocol Buffers compatible.

source